+++ packages/breadboard/src/harness/worker.ts
@@ -5,7 +5,7 @@
  */
 
 import type { HarnessRunResult, RunConfig } from "./types.js";
-import { RunState, asyncGen } from "../index.js";
+import { asyncGen } from "../index.js";
 import { createSecretAskingKit } from "./secrets.js";
 import { ProxyServer } from "../remote/proxy.js";
 import {
@@ -15,6 +15,7 @@ import {
 } from "../remote/worker.js";
 import { RunClient } from "../remote/run.js";
 import { InitClient } from "../remote/init.js";
+import type { RunState } from "../run/types.js";
 
 export const createWorker = (url: string) => {
   const workerURL = new URL(url, location.href);
