+++ packages/breadboard/src/remote/run.ts
@@ -8,6 +8,7 @@ import { Diagnostics } from "../harness/diagnostics.js";
 import { extractError } from "../harness/error.js";
 import { RunResult } from "../run.js";
 import { createRunStateManager } from "../run/index.js";
+import { RunState } from "../run/types.js";
 import { BoardRunner } from "../runner.js";
 import {
   WritableResult,
@@ -15,12 +16,7 @@ import {
   stubOutStreams,
 } from "../stream.js";
 import { timestamp } from "../timestamp.js";
-import {
-  InputValues,
-  NodeHandlerContext,
-  OutputValues,
-  RunState,
-} from "../types.js";
+import { InputValues, NodeHandlerContext, OutputValues } from "../types.js";
 import {
   AnyClientRunResult,
   AnyRunRequestMessage,
