+++ packages/breadboard/src/run.ts
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import type { RunState } from "./run/types.js";
 import { loadRunnerState, saveRunnerState } from "./serialization.js";
 import { timestamp } from "./timestamp.js";
 import type {
@@ -13,7 +14,6 @@ import type {
   TraversalResult,
   BreadboardRunResult,
   RunResultType,
-  RunState,
 } from "./types.js";
 
 export class RunResult implements BreadboardRunResult {
