+++ packages/breadboard/src/types.ts
@@ -22,6 +22,7 @@ import {
   InlineDataCapabilityPart,
   StoredDataCapabilityPart,
 } from "./data/types.js";
+import { ManagedRunState, RunState } from "./run/types.js";
 
 export type {
   Capability,
@@ -456,40 +457,6 @@ export interface BreadboardValidator {
   ): BreadboardValidator;
 }
 
-/**
- * Sequential number of the invocation of a node.
- * Useful for understanding the relative position of a
- * given invocation of node within the run.
- */
-export type InvocationId = number;
-
-/**
- * Information about a given invocation of a graph and
- * node within the graph.
- */
-export type RunStackEntry = {
-  /**
-   * The URL of the graph being run;
-   */
-  url: string | undefined;
-  /**
-   * The invocation id of the node within that graph.
-   */
-  node: InvocationId;
-  /**
-   * The state of the graph traversal at the time of the invocation.
-   */
-  state?: string;
-};
-
-/**
- * A stack of all invocations of graphs and nodes within the graphs.
- * The stack is ordered from the outermost graph to the innermost graph
- * that is currently being run.
- * Can be used to understand the current state of the run.
- */
-export type RunState = RunStackEntry[];
-
 export type GraphStartProbeData = {
   graph: GraphDescriptor;
   path: number[];
@@ -710,30 +677,6 @@ export type BreadboardCapability =
   | GraphDescriptorBoardCapability
   | ResolvedURLBoardCapability
   | UnresolvedPathBoardCapability;
-
-export type ManagedRunStateLifecycle = {
-  /**
-   * Signifies the beginning of a new graph being run or invoked.
-   * @param url -- url of the graph that is starting
-   */
-  dispatchGraphStart(url: string): void;
-  dispatchNodeStart(result: TraversalResult): void;
-  dispatchNodeEnd(): void;
-  dispatchGraphEnd(): void;
-  dispatchSkip(): void;
-  state(): Promise<RunState>;
-};
-
-/**
- * The representation of Breadboard runtime.
- * TODO: Rename to ManagedRuntime.
- */
-export type ManagedRunState = {
-  /**
-   * The entry point for signaling run lifecycle changes.
-   */
-  lifecycle(): ManagedRunStateLifecycle;
-};
 
 export interface NodeHandlerContext {
   readonly board?: GraphDescriptor;
